:recycle: Separate sleep animation from state
